backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
anthropic==0.42.0
tenacity==9.0.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
slack_sdk>=3.33.0
aiohttp>=3.9.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
